Open Addressed Hashing and Benchmarking
All uint32_t are changed to uin64_t

The large edits to Chained.h are just reformatting (removing tabs)

+++ Product.h
@@ -4,11 +4,11 @@
 
 class Product{
     public:
-        uint32_t id;
-        uint32_t quantity;
-        uint32_t price;
+        uint64_t id;
+        uint64_t quantity;
+        uint64_t price;
 
-        Product(uint32_t id, uint32_t quantity, uint32_t price) : 
+        Product(uint64_t id, uint64_t quantity, uint64_t price) :
             id(id),
             quantity(quantity), 
             price(price) { }
